Edit Workbook, Charlotte and Karen, 5 April, 2016
The transcript was difficult to read as it was only slightly edited. As a newcomer when first onboarding, Karen was really confused by this section, it turned her off from the book for nearly half a year. Given that it's an introduction to the project, written largely for newbies, we decided it might be more enticing if edited to be more like written English than spoken English